Update My orders and related functionality to act as modules
