+++ client/app/components/components.module.js
@@ -39,6 +39,7 @@ import { ModalActionsComponent } from './modal-actions/modal-actions.component.j
 import { BaseModalController } from './modal/base-modal-controller.js';
 import { BaseModalFactory } from './modal/base-modal.factory.js';
 import { NavigationController } from './navigation/navigation-contoller.js';
+import { OrderExplorerComponent } from './order-explorer/order-explorer.component.js';
 import { OwnershipServiceModalComponent } from './ownership-service-modal/ownership-service-modal.component.js';
 import { PaginationComponent } from './pagination/pagination.component.js';
 import { PowerStateComponent } from './power-state/power-state.component.js';
@@ -103,6 +104,7 @@ export default angular
   .factory('ModalService', BaseModalFactory)
   .controller('NavigationController', NavigationController)
   .component('ownershipServiceModal', OwnershipServiceModalComponent)
+  .component('orderExplorer', OrderExplorerComponent)
   .component('explorerPagination', PaginationComponent)
   .component('powerState', PowerStateComponent)
   .component('processRequestsModal', ProcessRequestsModalComponent)
